requirements.txt
altgraph==0.17
bcrypt==3.1.7
bump2version==1.0.0
bumpversion==0.6.0
cffi==1.14.0
click==7.1.2
cryptography==3.0
invoke==1.4.1
paramiko==2.7.1
pip-tools==5.3.1
pycparser==2.20
PyInstaller==3.6
PyNaCl==1.4.0
PyYAML==5.3.1
six==1.15.0
SQLAlchemy==1.3.18
xlwt==1.3.0
